WIP: podman library and tests
Most of the routines return a subprocess.CompletedProcess
which contains the .returncode, .stdout, and .stderr
attributes (the latter two are str).

You pull an image with `podman.pull()`, run a container
with `podman.run()`, etc.  `podman.run()` has keyword
arguments for some (but not all) common `podman run`
options.

Note that the default pull policy is 'never'; I think
for tests, pulls should be explicit (perhaps we should
even be loading from a tarball via `podman.load()`
instead of pulling from a registry).